refactor: change structure of gizmo files
